pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nomia"
version = "0.1.0"
description = "Business rule tracking for Python functions"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "PyYAML>=6.0.1"
]

[project.scripts]
nomia = "nomia.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["nomia*"]